Support java.lang.Iterable interface in QueryResultPage/ScanResultPage
